test: fixing build dependencies
